refactor(frontend): split TaxPlanningPage into tax-planning/ module
