+++ frontend/src/App.tsx
@@ -25,7 +25,7 @@ const pageImports = {
   InvestmentAnalyticsPage: () => import('@/pages/InvestmentAnalyticsPage'),
   MutualFundProjectionPage: () => import('@/pages/MutualFundProjectionPage'),
   ReturnsAnalysisPage: () => import('@/pages/ReturnsAnalysisPage'),
-  TaxPlanningPage: () => import('@/pages/TaxPlanningPage'),
+  TaxPlanningPage: () => import('@/pages/tax-planning/TaxPlanningPage'),
   GSTAnalysisPage: () => import('@/pages/GSTAnalysisPage'),
   NetWorthPage: () => import('@/pages/NetWorthPage'),
   SpendingAnalysisPage: () => import('@/pages/SpendingAnalysisPage'),
